lastite for arbitrary save

+++ src/pipeline.py
@@ -948,7 +948,7 @@ class Plots:
                 os.remove(f'figures_{self.job_id}/gain_iter{self._steps}.png')
 
             plt.close()
-                       
+
 class Pipeline(Chi2, Plots):
 
 
@@ -1120,26 +1120,34 @@ class Pipeline(Chi2, Plots):
         if self.rank == 0:
             folder = self.params['foldername'] + f"_seed{str(self.params['CMB']['seed'])}"
             if self.params['savelastite']:
- 	        for filename in np.sort(glob.glob(folder+'/'+self.params['filename']+'*.pkl')):
-		    try:
-		        os.remove(filename)
-	            except OSError:
-		        pass
+                for filename in np.sort(glob.glob(folder+'/'+self.params['filename']+'*.pkl')):
+                    try:
+                        os.remove(filename)
+                    except OSError:
+                        pass
+                    
                 with open(folder + '/' + self.params['filename']+f"_{self._steps+1}_{str(self.params['CMB']['iter'])}.pkl", 'wb') as handle:
-                    pickle.dump({'components':self.components, 
-                         	'components_i':self.components_iter,
-                         	'beta':self.beta,
-                         	'g':self.g,
-                         	'gi':self.g_iter,
-                         	'allg':self.allg,
-                         	'G':self.G,
-                         	'center':self.center,
-                         	'coverage':self.coverage,
-                         	'seenpix':self.seenpix}, handle, protocol=pickle.HIGHEST_PROTOCOL)
-            if self.params['save'] != 0:
-                if (self._steps+1) % self.params['save'] == 0:                   
-                    with open(folder + '/' + self.params['filename']+f"_{self._steps+1}_{str(self.params['CMB']['iter'])}.pkl", 'wb') as handle:
-                        pickle.dump({'components':self.components, 
+                    pickle.dump({'components':self.components,
+                                 'components_i':self.components_iter,
+                                 'beta':self.beta,
+                                 'g':self.g,
+                                 'gi':self.g_iter,
+                                 'allg':self.allg,
+                                 'G':self.G,
+                                 'center':self.center,
+                                 'coverage':self.coverage,
+                                 'seenpix':self.seenpix}, handle, protocol=pickle.HIGHEST_PROTOCOL)
+            else:
+                if self.params['save'] != 0:
+                    if (self._steps+1) % self.params['save'] == 0:
+                        if self.params['savelastite']:
+                            for filename in np.sort(glob.glob(folder+'/'+self.params['filename']+'*.pkl')):
+                                try:
+                                    os.remove(filename)
+                                except OSError:
+                                    pass
+                        with open(folder + '/' + self.params['filename']+f"_{self._steps+1}_{str(self.params['CMB']['iter'])}.pkl", 'wb') as handle:
+                            pickle.dump({'components':self.components, 
                                  'components_i':self.components_iter,
                                  'beta':self.beta,
                                  'g':self.g,
